t:wqRevert "split up FlatteningConfigsTest"
This reverts commit 3da1e5b4aa209b7aa85aabcf50fbed1940866945.
